SP06 - Call render video from video info view

+++ video_generator/routers/api_router.js
@@ -33,34 +33,50 @@ router.post('/add-frame', function (req, res) {
 
 
 router.post('/render', function (req, res) {
-    var imgDir = sprintf('%s/%s', preRenderDir, req.body.filename);
-    var fullFilePath = sprintf('%s/%s.mp4', videoDir, req.body.filename);
+    var videoId = req.body.video_id;
 
-    fs_extra.ensureDirSync(videoDir);
-    fs_extra.removeSync(fullFilePath);
+    if (videoId) {
+        var imgDir = sprintf('%s/%s', preRenderDir, videoId);
+        var fullFilePath = sprintf('%s/%s.mp4', videoDir, videoId);
 
-    console.log("Begining rendering of your video. This might take a long time...")
-    var ffmpeg = cp.spawn('ffmpeg', [
-        '-framerate', '24',
-        '-start_number', '0',
-        '-i', 'image-%010d.png',
-        '-refs', '5',
-        '-c:v', 'libx264',
-        '-preset', 'fast',
-        '-pix_fmt', 'yuv420p',
-        '-crf', '18',
-        fullFilePath
-    ], {
-        cwd: imgDir,
-        stdio: 'inherit'
-    });
+        fs_extra.ensureDirSync(videoDir);
+        fs_extra.removeSync(fullFilePath);
 
-    ffmpeg.on('close', function (code) {
-        console.log(sprintf('Finished rendering video. You can find it at %s/%s.mp4', videoDir, req.body.filename));
-        fs_extra.removeSync(imgDir);
-    });
+        console.log("Begining rendering of your video. This might take a long time...")
+        var ffmpeg = cp.spawn('ffmpeg', [
+            '-framerate', '24',
+            '-start_number', '0',
+            '-i', 'image-%010d.png',
+            '-refs', '5',
+            '-c:v', 'libx264',
+            '-preset', 'fast',
+            '-pix_fmt', 'yuv420p',
+            '-crf', '18',
+            fullFilePath
+        ], {
+            cwd: imgDir,
+            stdio: 'inherit'
+        });
 
-    res.end();
+        ffmpeg.on('close', function (code) {
+            console.log(sprintf('Finished rendering video. You can find it at %s/%s.mp4', videoDir, videoId));
+            fs_extra.removeSync(imgDir);
+
+            dal.updateVideoInfos(
+                videoId, {
+                    status: 'RENDERED'
+                },
+                function (err, data) {
+                    if (err) console.error(error);
+                    console.log(data);
+                }
+            );
+            
+            res.end();
+        });
+    } else {
+        res.end();
+    }
 });
